Add files via upload
"If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3,5,6 and 9. The sum of these multiples is 23. 
Find the sum of all the multiples of 3 or 5 below 1000."
this a solution to this problem in python. its not perfect though. It wont work for any numbers where their product isn't their LCM. to complete this, I need to find a way to calculate the LCM of two numbers. 
This solution is based on existing mathematics. Specifically, natural summation formula. Any have a way to find the LCM of 2 numbers in an efficient way.